{"schema":"decodex/commit/1","summary":"Move macOS passive shell runtime","authority":"manual"}
